requirements.txt
langchain-openai
langgraph
pydantic
docker
python-dotenv
colorama
termcolor
aiofiles
